- code_warning_fix: Removed an 64 bit compiler warning. (Thanks stahta01)
Warning happened when doing 64 bit compile on Windows under MSys2.

+++ src/plugins/compilergcc/depslib/src/filent.c
@@ -280,7 +280,7 @@ file_archscan(
 	    if( ( c = strrchr( name, '\\' ) ) ) /* TNB */
 		name = c + 1;
 
-	    sprintf( buf, "%s(%.*s)", archive, endname - name, name );
+	    sprintf( buf, "%s(%.*s)", archive, (int)(endname - name), name );
 	    (*func)( closure, buf, 1 /* time valid */, (time_t)lar_date );
 
 	    offset += SARHDR + lar_size;
